Randomize discover when clicking between home and nav
